Apply documentation changes post- artifacts refactor
